2015 fallow field daily estimates of ET
In this .csv, the daily ET estimates are listed for the five stations deployed in fallow fields in September 2015.  
Station D5 had sensors allowing two methods of estimating ET, so there are two columns for that station; one column listing ET estimated from the surface renewal method and one column listing ET estimated from the eddy covariance method.  
Station D4 had a net radiometer that did not yield accurate or precise results in a post-deployment field calibration, so due to that sensor's slightly erroneous data recording, we chose not to include the 2015 field data from D4 in our field team's analysis of ET estimates and advise that the field D4 data also not be used in comparison analysis.